requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.0
crypto==1.4.1
cryptography==46.0.5
cycler==0.12.1
Django==4.2.28
exceptiongroup==1.3.1
faiss-cpu==1.13.0
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
grad-cam==1.5.5
gunicorn==23.0.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.7
kombu==5.6.2
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
Naked==0.1.32
networkx==3.2.1
numpy==2.0.2
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
polars==1.36.1
polars-runtime-32==1.36.1
prompt_toolkit==3.0.52
psutil==7.2.2
pycparser==2.23
pycryptodome==3.23.0
pyparsing==3.3.2
pypdfium2==5.5.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
shellescape==3.8.1
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.3
ttach==0.0.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ultralytics==8.4.19
ultralytics-thop==2.0.18
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
xgboost==2.1.4
zipp==3.23.0
